requirements.txt
# API Server Dependencies
fastapi==0.109.0
uvicorn>=0.31.1
motor==3.6.0  # Async MongoDB driver
pydantic>=2.5.3
pydantic-settings>=2.5.2
email-validator>=2.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
python-multipart>=0.0.9

# MCP Server Dependencies
mcp==1.26.0
httpx>=0.27.1

# Groq API
groq>=0.11.0

# UI Dependencies
streamlit==1.30.0

# Development Dependencies
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
